requirements.txt
# VibeQuant Dependencies

# Core - Required for all modes
alpaca-py>=0.13.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyarrow>=14.0.0  # For parquet caching

# Autonomous mode only (not needed for --interactive)
# langgraph>=0.0.40
# langchain>=0.1.0
# langchain-openai>=0.0.5
# langchain-anthropic>=0.1.0
